build errors

+++ Types.fs
@@ -1,8 +1,8 @@
 module Types
 type Payment = {
-    event : string
-    transaction_id : string
-    amount : decimal
-    currency : string
-    timestamp : string
+    event: string
+    transaction_id: string
+    amount: float
+    currency: string
+    timestamp: string
 }
